failai pravaryti per indent; oled ekranelis su u8g2 biblioteka

+++ inc/hd44780.h
@@ -5,15 +5,15 @@ void HD44780_InitPins(void);
 void HD44780_Init(void);
 void HD44780_ClearDisplay(void);
 void HD44780_ReturnHome(void);
-void HD44780_EntryModeSet(char,char);
-void HD44780_DisplayControl(char,char,char);
-void HD44780_Shift(char,char);
-void HD44780_FunctionSet(char,char,char);
+void HD44780_EntryModeSet(char, char);
+void HD44780_DisplayControl(char, char, char);
+void HD44780_Shift(char, char);
+void HD44780_FunctionSet(char, char, char);
 void HD44780_SetCGRAM(unsigned char);
 void HD44780_SetDDRAM(unsigned char);
-void HD44780_CommandWrite(unsigned char,unsigned short);
+void HD44780_CommandWrite(unsigned char,int);
 void HD44780_DataWrite(unsigned char);
-void HD44780_DisplayString(char*,int,int,int);
+void HD44780_DisplayString(char *, int, int, int);
 void HD44780_ClearLine(int);
 
 #endif
